+++ src/fmalloc_internal.h
@@ -55,10 +55,11 @@ struct fm_vector {
     uint64_t catalog_offset;
     uint64_t generation;
     SEXP refs;
+    size_t parent_refs;
 
     fm_vector(fm_runtime *_runtime, SEXPTYPE _type, R_xlen_t _length, void *_data, size_t _bytes)
         : runtime(_runtime), type(_type), len(_length), data(_data), bytes(_bytes),
-          catalog_offset(0), generation(0), refs(R_NilValue) {}
+          catalog_offset(0), generation(0), refs(R_NilValue), parent_refs(0) {}
 };
 
 static constexpr uint32_t FM_STRING_FLAG_NA = 1u;